Convert complex representations to two floats instead of doubles, update state and gate representations accordingly.

+++ src/main/scala/scotty/quantum/QuantumContext.scala
@@ -29,8 +29,8 @@ trait QuantumContext {
 }
 
 object QuantumContext {
-  type Vector = Array[Double]
-  type Matrix = Array[Array[Double]]
+  type Vector = Array[Float]
+  type Matrix = Array[Array[Float]]
 
   case class QuantumException(message: String) extends Exception(message)
 }
